move helper func to utils module
